Move Opcode/OpcodeInfo to opcode.{c,h}
I forgot to set s_initialized to true in init_opcode_info, which meant
that it was _really_ slow (on my laptop, parsing AngryBots.wast, 75% of
the time was spent in init_opcode_info). Whoops.

I could have fixed it by just setting s_initialized to true, but instead
I just filled in the gaps with operations called "Invalid_0x16" etc.

I also took this as an opportunity to move the opcode stuff into its own
files.
