+++ CMakeLists.txt
@@ -212,6 +212,7 @@ endif ()
 add_custom_target(everything)
 
 add_library(libwabt STATIC
+  src/opcode.cc
   src/binary-error-handler.cc
   src/source-error-handler.cc
   src/hash-util.cc
